+++ Scripts/Attributes/ButtonAttribute.cs
@@ -19,7 +19,7 @@ namespace EditorAttributes
 		/// </summary>
 		/// <param name="buttonLabel">The label displayed on the button</param>
 		/// <param name="buttonHeight">The height of the button</param>
-		public ButtonAttribute(string buttonLabel = "", float buttonHeight = 17f)
+		public ButtonAttribute(string buttonLabel = "", float buttonHeight = 18f)
 		{
 			ConditionName = "";
 			ButtonLabel = buttonLabel;
@@ -30,11 +30,11 @@ namespace EditorAttributes
 		/// Attribute to add a button in the inspector
 		/// </summary>
 		/// <param name="conditionName">The name of the condition to evaluate</param>
-		/// <param name="conditionResult">What happens to the property when the condition evaluates to true</param>
+		/// <param name="conditionResult">What happens to the button when the condition evaluates to true</param>
 		/// <param name="negate">Negate the evaluated condition</param>
 		/// <param name="buttonLabel">The label displayed on the button</param>
 		/// <param name="buttonHeight">The height of the button</param>
-		public ButtonAttribute(string conditionName, ConditionResult conditionResult, bool negate = false, string buttonLabel = "", float buttonHeight = 17f) : this(buttonLabel, buttonHeight)
+		public ButtonAttribute(string conditionName, ConditionResult conditionResult, bool negate = false, string buttonLabel = "", float buttonHeight = 18f) : this(buttonLabel, buttonHeight)
 		{
 			ConditionResult = conditionResult;
 			ConditionName = conditionName;
@@ -46,11 +46,11 @@ namespace EditorAttributes
 		/// </summary>
 		/// <param name="conditionName">The name of the condition to evaluate</param>
 		/// <param name="enumValue">The value of the enum condition</param>
-		/// <param name="conditionResult">What happens to the property when the condition evaluates to true</param>
+		/// <param name="conditionResult">What happens to the button when the condition evaluates to true</param>
 		/// <param name="negate">Negate the evaluated condition</param>
 		/// <param name="buttonLabel">The label displayed on the button</param>
 		/// <param name="buttonHeight">The height of the button</param>
-		public ButtonAttribute(string conditionName, object enumValue, ConditionResult conditionResult, bool negate = false, string buttonLabel = "", float buttonHeight = 17f) : this(buttonLabel, buttonHeight)
+		public ButtonAttribute(string conditionName, object enumValue, ConditionResult conditionResult, bool negate = false, string buttonLabel = "", float buttonHeight = 18f) : this(buttonLabel, buttonHeight)
 		{
 			ConditionResult = conditionResult;
 			ConditionName = conditionName;