+++ Scripts/Attributes/Editor/DisableFieldDrawer.cs
@@ -1,21 +1,17 @@
 using UnityEditor;
 using UnityEngine;
-using System.Reflection;
 
 namespace EditorAttributes.Editor
 {
     [CustomPropertyDrawer(typeof(DisableFieldAttribute))]
     public class DisableFieldDrawer : PropertyDrawerBase
     {
-		private MemberInfo conditionalProperty;
-
 		public override void OnGUI(Rect position, SerializedProperty property, GUIContent label)
 		{
 			var disableAttribute = attribute as DisableFieldAttribute;
+			var conditionalProperty = GetValidMemberInfo(disableAttribute.ConditionName, property);
 
-			conditionalProperty = GetValidMemberInfo(disableAttribute.ConditionName, property);
-
-			GUI.enabled = !GetConditionValue<DisableFieldAttribute>(conditionalProperty, attribute, property.serializedObject.targetObject, true);
+			GUI.enabled = !GetConditionValue<DisableFieldAttribute>(conditionalProperty, attribute, property.serializedObject.targetObject);
 
 			DrawProperty(position, property, label);