+++ Scripts/Attributes/Editor/EnableFieldDrawer.cs
@@ -1,21 +1,17 @@
 using UnityEditor;
 using UnityEngine;
-using System.Reflection;
 
 namespace EditorAttributes.Editor
 {
     [CustomPropertyDrawer(typeof(EnableFieldAttribute))]
     public class EnableFieldDrawer : PropertyDrawerBase
     {
-		private MemberInfo conditionalProperty;
-
 		public override void OnGUI(Rect position, SerializedProperty property, GUIContent label)
 		{
 			var enableAttribute = attribute as EnableFieldAttribute;
+			var conditionalProperty = GetValidMemberInfo(enableAttribute.ConditionName, property);
 
-			conditionalProperty = GetValidMemberInfo(enableAttribute.ConditionName, property);
-
-			GUI.enabled = GetConditionValue<EnableFieldAttribute>(conditionalProperty, attribute, property.serializedObject.targetObject, true);
+			GUI.enabled = GetConditionValue<EnableFieldAttribute>(conditionalProperty, attribute, property.serializedObject.targetObject);
 			
 			DrawProperty(position, property, label);