+++ Scripts/Attributes/Editor/MessageBoxDrawer.cs
@@ -1,21 +1,17 @@
 using UnityEditor;
 using UnityEngine;
-using System.Reflection;
 
 namespace EditorAttributes.Editor
 {
     [CustomPropertyDrawer(typeof(MessageBoxAttribute))]
     public class MessageBoxDrawer : PropertyDrawerBase
     {
-		private MemberInfo conditionalProperty;
-
 		public override void OnGUI(Rect position, SerializedProperty property, GUIContent label)
 		{
 			var messageBox = attribute as MessageBoxAttribute;
+			var conditionalProperty = GetValidMemberInfo(messageBox.ConditionName, property);
 
-			conditionalProperty = GetValidMemberInfo(messageBox.ConditionName, property);
-
-			if (GetConditionValue<MessageBoxAttribute>(conditionalProperty, attribute, property.serializedObject.targetObject, true)) EditorGUILayout.HelpBox(messageBox.Message, (MessageType)messageBox.MessageType);
+			if (GetConditionValue<MessageBoxAttribute>(conditionalProperty, attribute, property.serializedObject.targetObject)) EditorGUILayout.HelpBox(messageBox.Message, (MessageType)messageBox.MessageType);
 
 			if (messageBox.DrawProperty) DrawProperty(position, property, label);
 		}