+++ Scripts/Attributes/GUIColorAttribute.cs
@@ -7,6 +7,7 @@ namespace EditorAttributes
 		public float R { get; private set; }
 		public float G { get; private set; }
 		public float B { get; private set; }
+		public bool UseRGB { get; private set; }
 		public string HexColor { get; private set; }
 
 		public GUIColor GUIColor { get; private set; }
@@ -15,7 +16,7 @@ namespace EditorAttributes
 		/// Attribute to color the GUI
 		/// </summary>
 		/// <param name="fieldColor">The color of the GUI</param>
-	    public GUIColorAttribute(GUIColor fieldColor) => GUIColor = fieldColor;
+		public GUIColorAttribute(GUIColor fieldColor) => GUIColor = fieldColor;
 
 		/// <summary>
 		/// Attribute to color the GUI
@@ -25,7 +26,7 @@ namespace EditorAttributes
 		/// <param name="b">Blue amount</param>
 		public GUIColorAttribute(float r, float g, float b)
 		{
-			GUIColor = GUIColor.UseRGB;
+			UseRGB = true;
 			R = r;
 			G = g;
 			B = b;