+++ Scripts/Attributes/IColorAttribute.cs
@@ -5,6 +5,7 @@ namespace EditorAttributes
 		public float R { get; }
 		public float G { get; }
 		public float B { get; }
+		public bool UseRGB { get; }
 		public string HexColor { get; }
 
 		public GUIColor GUIColor { get; }